+++ src/diffs/gfa.rs
@@ -419,7 +419,7 @@ mod tests {
         let new_child_block_group = Collection::get_block_groups(conn, "test collection 2", None)
             .pop()
             .unwrap();
-        let all_child_sequences = gen_models_graph_tests::get_all_sequences_with_pruning(
+        let all_child_sequences = gen_graph::models::get_all_sequences_with_pruning(
             conn,
             &new_child_block_group.id,
             false,
@@ -508,7 +508,7 @@ mod tests {
             Collection::get_block_groups(conn, "test collection 3", None)
                 .pop()
                 .unwrap();
-        let all_grandchild_sequences = gen_models_graph_tests::get_all_sequences_with_pruning(
+        let all_grandchild_sequences = gen_graph::models::get_all_sequences_with_pruning(
             conn,
             &new_grandchild_block_group.id,
             false,
@@ -543,7 +543,7 @@ mod tests {
             Collection::get_block_groups(conn, "test collection 4", None)
                 .pop()
                 .unwrap();
-        let all_grandchild_sequences = gen_models_graph_tests::get_all_sequences_with_pruning(
+        let all_grandchild_sequences = gen_graph::models::get_all_sequences_with_pruning(
             conn,
             &new_grandchild_block_group.id,
             false,
@@ -654,12 +654,9 @@ mod tests {
         let new_block_group = Collection::get_block_groups(conn, "test collection 2", None)
             .pop()
             .unwrap();
-        let all_sequences = gen_models_graph_tests::get_all_sequences_with_pruning(
-            conn,
-            &new_block_group.id,
-            false,
-        )
-        .unwrap();
+        let all_sequences =
+            gen_graph::models::get_all_sequences_with_pruning(conn, &new_block_group.id, false)
+                .unwrap();
 
         assert_eq!(
             all_sequences,
@@ -765,12 +762,9 @@ mod tests {
         let new_block_group = Collection::get_block_groups(conn, "test collection 2", None)
             .pop()
             .unwrap();
-        let all_sequences = gen_models_graph_tests::get_all_sequences_with_pruning(
-            conn,
-            &new_block_group.id,
-            false,
-        )
-        .unwrap();
+        let all_sequences =
+            gen_graph::models::get_all_sequences_with_pruning(conn, &new_block_group.id, false)
+                .unwrap();
 
         assert_eq!(
             all_sequences,
@@ -935,12 +929,9 @@ mod tests {
         let new_block_group = Collection::get_block_groups(conn, "test collection 3", None)
             .pop()
             .unwrap();
-        let all_sequences = gen_models_graph_tests::get_all_sequences_with_pruning(
-            conn,
-            &new_block_group.id,
-            false,
-        )
-        .unwrap();
+        let all_sequences =
+            gen_graph::models::get_all_sequences_with_pruning(conn, &new_block_group.id, false)
+                .unwrap();
 
         assert_eq!(
             all_sequences,
@@ -1106,12 +1097,9 @@ mod tests {
         let new_block_group = Collection::get_block_groups(conn, "test collection 3", None)
             .pop()
             .unwrap();
-        let all_sequences = gen_models_graph_tests::get_all_sequences_with_pruning(
-            conn,
-            &new_block_group.id,
-            false,
-        )
-        .unwrap();
+        let all_sequences =
+            gen_graph::models::get_all_sequences_with_pruning(conn, &new_block_group.id, false)
+                .unwrap();
 
         assert_eq!(
             all_sequences,
@@ -1253,7 +1241,7 @@ mod tests {
         let new_child_block_group = Collection::get_block_groups(conn, "test collection 2", None)
             .pop()
             .unwrap();
-        let all_child_sequences = gen_models_graph_tests::get_all_sequences_with_pruning(
+        let all_child_sequences = gen_graph::models::get_all_sequences_with_pruning(
             conn,
             &new_child_block_group.id,
             false,
@@ -1343,7 +1331,7 @@ mod tests {
             Collection::get_block_groups(conn, "test collection 3", None)
                 .pop()
                 .unwrap();
-        let all_grandchild_sequences = gen_models_graph_tests::get_all_sequences_with_pruning(
+        let all_grandchild_sequences = gen_graph::models::get_all_sequences_with_pruning(
             conn,
             &new_grandchild_block_group.id,
             false,
@@ -1375,7 +1363,7 @@ mod tests {
             Collection::get_block_groups(conn, "test collection 4", None)
                 .pop()
                 .unwrap();
-        let all_grandchild_sequences = gen_models_graph_tests::get_all_sequences_with_pruning(
+        let all_grandchild_sequences = gen_graph::models::get_all_sequences_with_pruning(
             conn,
             &new_grandchild_block_group.id,
             false,
