+++ src/graphs/translation.rs
@@ -1864,8 +1864,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         let protein =
             translate_annotation(&conn, &annotation, Some(&block_group_id), params).unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["ME*".to_string()])
         );
     }
@@ -1913,8 +1912,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         let protein =
             translate_annotation(&conn, &annotation, Some(&block_group_id), params).unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["GM".to_string()])
         );
     }
@@ -1928,8 +1926,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         let protein =
             translate_annotation(&conn, &annotation, Some(&block_group_id), params).unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["ME*".to_string()])
         );
     }
@@ -1943,8 +1940,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         let protein =
             translate_annotation(&conn, &annotation, Some(&block_group_id), params).unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["ME*".to_string()])
         );
     }
@@ -1958,8 +1954,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         let protein =
             translate_annotation(&conn, &annotation, Some(&block_group_id), params).unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["WE*".to_string()])
         );
     }
@@ -1972,8 +1967,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         let protein =
             translate_annotation(&conn, &annotation, Some(&block_group_id), params).unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["MW".to_string()])
         );
     }
@@ -2065,8 +2059,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
             translate_annotation(&conn, &annotation, Some(&block_group.id), params).unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from([expected.to_string()]),
             "reverse-strand protein should read N → C left to right"
         );
@@ -2243,8 +2236,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         let protein =
             translate_annotation(&conn, &annotation, Some(&block_group_id), params).unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["M*".to_string()])
         );
     }
@@ -2305,8 +2297,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         )
         .unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["ME*".to_string()]),
             "translation should only follow the annotation's own entry node"
         );
@@ -2322,8 +2313,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         let protein =
             translate_annotation(&conn, &annotation, Some(&block_group_id), params).unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["*".to_string()])
         );
     }
@@ -2344,8 +2334,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         let protein =
             translate_annotation(&conn, &annotation, Some(&block_group_id), params).unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["MP*".to_string(), "MPVG*".to_string()]),
             "frameshift deletion should truncate the variant protein at the premature stop"
         );
@@ -2366,8 +2355,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         let protein =
             translate_annotation(&conn, &annotation, Some(&block_group_id), params).unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["MK*".to_string(), "MND*".to_string()]),
             "frameshift deletion should read through the wild-type stop to a later one"
         );
@@ -2559,14 +2547,14 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         .unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &parent_protein.id, true)
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &parent_protein.id, true)
                 .unwrap(),
             HashSet::from(["ME*".to_string()])
         );
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &child_protein.id, true)
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &child_protein.id, true)
                 .unwrap(),
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &parent_protein.id, true)
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &parent_protein.id, true)
                 .unwrap(),
             "upstream deletion changed the annotated protein",
         );
@@ -2644,19 +2632,19 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         .unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &parent_protein.id, true)
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &parent_protein.id, true)
                 .unwrap(),
             HashSet::from(["ME*".to_string()])
         );
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &child_protein.id, true)
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &child_protein.id, true)
                 .unwrap(),
             HashSet::from(["MQ*".to_string()])
         );
         assert_ne!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &child_protein.id, true)
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &child_protein.id, true)
                 .unwrap(),
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &parent_protein.id, true)
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &parent_protein.id, true)
                 .unwrap(),
             "point mutation inside the CDS did not change the protein",
         );
@@ -2723,8 +2711,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         .unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["ME*".to_string()]),
             "translation should only follow the annotation's own entry node"
         );
@@ -2789,8 +2776,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         .unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["ME*".to_string(), "MEW".to_string()]),
             "last-base variant was dropped from the extracted subgraph",
         );
@@ -2843,8 +2829,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
             translate_from_path(&conn, &block_group.id, 0, TranslationParams::new("test")).unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &from_path.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &from_path.id, true).unwrap(),
             HashSet::from(["*".to_string()]),
             "translate_from_path(0) should only translate the literal entry node (A), \
              not the unrelated D/P/Z branch reachable from PATH_START"
@@ -2866,8 +2851,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         let protein = translate_block_group(&conn, &block_group_id, params).unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["MKPEGF*".to_string(), "MKPQGF*".to_string()])
         );
     }
@@ -2893,8 +2877,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
             translate_annotation(&conn, &annotation, Some(&block_group_id), params).unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["MKPEGF*".to_string(), "MKPQGF*".to_string()])
         );
     }
@@ -2911,8 +2894,7 @@ ncbieaa  "FFLLSSSSYY**CC*WLLLLPPPPHHQQRRRRIIIMTTTTNNKKSSRRVVVVAAAADDEEGGGG"
         let protein = translate_from_path(&conn, &block_group_id, 0, params).unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &protein.id, true)
-                .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &protein.id, true).unwrap(),
             HashSet::from(["MKPEGF*".to_string(), "MKPQGF*".to_string()])
         );
     }
