+++ src/imports/fasta.rs
@@ -243,7 +243,7 @@ mod tests {
 
         let block_group_id = BlockGroup::get_id("test", Sample::DEFAULT_NAME, "m123", None);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(conn, &block_group_id, false)
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_group_id, false)
                 .unwrap(),
             HashSet::from_iter(vec!["ATCGATCGATCGATCGATCGGGAACACACAGAGA".to_string()])
         );
@@ -273,7 +273,7 @@ mod tests {
         .unwrap();
         let block_group_id = BlockGroup::get_id("test", Sample::DEFAULT_NAME, "m123", None);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(conn, &block_group_id, false)
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_group_id, false)
                 .unwrap(),
             HashSet::from_iter(vec!["ATCGATCGATCGATCGATCGGGAACACACAGAGA".to_string()])
         );
@@ -321,7 +321,7 @@ mod tests {
         .unwrap();
         let block_group_id = BlockGroup::get_id("test", Sample::DEFAULT_NAME, "m123", None);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(conn, &block_group_id, false)
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_group_id, false)
                 .unwrap(),
             HashSet::from_iter(vec!["ATCGATCGATCGATCGATCGGGAACACACAGAGA".to_string()])
         );
@@ -345,7 +345,7 @@ mod tests {
         .unwrap();
         let block_group_id = BlockGroup::get_id("test", "new-sample", "m123", None);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(conn, &block_group_id, false)
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_group_id, false)
                 .unwrap(),
             HashSet::from_iter(vec!["ATCGATCGATCGATCGATCGGGAACACACAGAGA".to_string()])
         );
@@ -379,7 +379,7 @@ mod tests {
         .unwrap();
         let block_group_id = BlockGroup::get_id("test", Sample::DEFAULT_NAME, "m123", None);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(conn, &block_group_id, false)
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_group_id, false)
                 .unwrap(),
             HashSet::from_iter(vec!["ATCGATCGATCGATCGATCGGGAACACACAGAGA".to_string()])
         );
