+++ src/imports/genbank.rs
@@ -1199,12 +1199,9 @@ mod tests {
             let f = reader::parse_file(&path).unwrap();
             let seq = str::from_utf8(&f[0].seq).unwrap().to_string();
             let block_group_id = BlockGroup::get_id("", Sample::DEFAULT_NAME, "insertion", None);
-            let seqs = gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_group_id,
-                false,
-            )
-            .unwrap();
+            let seqs =
+                gen_graph::models::get_all_sequences_with_pruning(conn, &block_group_id, false)
+                    .unwrap();
             assert_eq!(
                 seqs,
                 HashSet::from_iter([
@@ -1252,12 +1249,9 @@ mod tests {
         GATGCCATTGGGATATATCAACGGTGGTATATCCAGTGATTTTTTTCTCCAT",
             );
             let block_group_id = BlockGroup::get_id("", Sample::DEFAULT_NAME, "deletion", None);
-            let seqs = gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_group_id,
-                false,
-            )
-            .unwrap();
+            let seqs =
+                gen_graph::models::get_all_sequences_with_pruning(conn, &block_group_id, false)
+                    .unwrap();
             assert_eq!(
                 seqs,
                 HashSet::from_iter([
@@ -1304,7 +1298,7 @@ mod tests {
              GGTACATTGAGCAACTGACTGAAATGCCTCAAAATGTTCTTTACGATGCCATTGGGAT
              ATATCAACGGTGGTATATCCAGTGATTTTTTTCTC",
             );
-            let seqs = gen_models_graph_tests::get_all_sequences_with_pruning(
+            let seqs = gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &BlockGroup::get_id("", Sample::DEFAULT_NAME, "deletion_and_insertion", None),
                 false,
@@ -1358,7 +1352,7 @@ mod tests {
              GGTACATTGAGCAACTGACTGAAATGCCTCAAAATGTTCTTTACGATGCCATTGGGAT
              ATATCAACGGTGGTATATCCAGTGATTTTTTTCTC",
             );
-            let seqs = gen_models_graph_tests::get_all_sequences_with_pruning(
+            let seqs = gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &BlockGroup::get_id("", Sample::DEFAULT_NAME, "substitution", None),
                 false,
@@ -1397,16 +1391,15 @@ mod tests {
             // there would be 4! sequences so we just check we have the fully changed and unchanged sequence
             let f = reader::parse_file(&path).unwrap();
             let mod_seq = str::from_utf8(&f[0].seq).unwrap().to_string();
-            let sequences: HashSet<String> =
-                gen_models_graph_tests::get_all_sequences_with_pruning(
-                    conn,
-                    &BlockGroup::get_id("", Sample::DEFAULT_NAME, "insertion", None),
-                    false,
-                )
-                .unwrap()
-                .iter()
-                .map(|s| s.to_lowercase())
-                .collect();
+            let sequences: HashSet<String> = gen_graph::models::get_all_sequences_with_pruning(
+                conn,
+                &BlockGroup::get_id("", Sample::DEFAULT_NAME, "insertion", None),
+                false,
+            )
+            .unwrap()
+            .iter()
+            .map(|s| s.to_lowercase())
+            .collect();
             let unchanged_seq = get_unmodified_sequence();
             assert!(sequences.contains(&mod_seq));
             assert!(sequences.contains(&unchanged_seq));
