+++ src/updates/fasta.rs
@@ -312,12 +312,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -426,12 +422,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -498,12 +490,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -570,12 +558,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -648,12 +632,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -720,12 +700,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -790,12 +766,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -847,7 +819,7 @@ mod tests {
         assert!(result.is_ok(), "{result:?}");
         let block_group = get_sample_bg(conn, &collection, "derived");
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(conn, &block_group.id, false)
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_group.id, false)
                 .unwrap(),
             HashSet::from_iter([
                 "ATCGATCGATCGATCGATCGGGAACACACAGAGA".to_string(),
@@ -902,12 +874,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
 
@@ -967,12 +935,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -1026,12 +990,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
