+++ src/updates/genbank.rs
@@ -380,15 +380,11 @@ mod tests {
             let mod_seq = str::from_utf8(&f[0].seq).unwrap().to_string();
             let block_group_id = BlockGroup::get_id("", Sample::DEFAULT_NAME, "insertion", None);
             let sequences: HashSet<String> =
-                gen_models_graph_tests::get_all_sequences_with_pruning(
-                    conn,
-                    &block_group_id,
-                    false,
-                )
-                .unwrap()
-                .iter()
-                .map(|s| s.to_lowercase())
-                .collect();
+                gen_graph::models::get_all_sequences_with_pruning(conn, &block_group_id, false)
+                    .unwrap()
+                    .iter()
+                    .map(|s| s.to_lowercase())
+                    .collect();
             let unchanged_seq = get_unmodified_sequence();
             assert!(sequences.contains(&mod_seq));
             assert!(sequences.contains(&unchanged_seq));
@@ -437,15 +433,11 @@ mod tests {
             let f = reader::parse_file(&path).unwrap();
             let block_group_id = BlockGroup::get_id("", Sample::DEFAULT_NAME, "insertion", None);
             let sequences: HashSet<String> =
-                gen_models_graph_tests::get_all_sequences_with_pruning(
-                    conn,
-                    &block_group_id,
-                    false,
-                )
-                .unwrap()
-                .iter()
-                .map(|s| s.to_lowercase())
-                .collect();
+                gen_graph::models::get_all_sequences_with_pruning(conn, &block_group_id, false)
+                    .unwrap()
+                    .iter()
+                    .map(|s| s.to_lowercase())
+                    .collect();
             let unchanged_seq = get_unmodified_sequence();
             assert!(sequences.contains(&unchanged_seq));
             let mod_seq = str::from_utf8(&f[0].seq).unwrap().to_string();
@@ -455,15 +447,11 @@ mod tests {
             // has a deletion in it.
             let block_group_id = BlockGroup::get_id("", Sample::DEFAULT_NAME, "deletion", None);
             let sequences: HashSet<String> =
-                gen_models_graph_tests::get_all_sequences_with_pruning(
-                    conn,
-                    &block_group_id,
-                    false,
-                )
-                .unwrap()
-                .iter()
-                .map(|s| s.to_lowercase())
-                .collect();
+                gen_graph::models::get_all_sequences_with_pruning(conn, &block_group_id, false)
+                    .unwrap()
+                    .iter()
+                    .map(|s| s.to_lowercase())
+                    .collect();
             let mod_seq = str::from_utf8(&f[1].seq).unwrap().to_string();
             assert!(sequences.contains(&unchanged_seq));
             assert!(sequences.contains(&mod_seq));
