+++ src/updates/gfa.rs
@@ -546,12 +546,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -601,12 +597,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
