+++ src/updates/sequence.rs
@@ -252,7 +252,7 @@ mod tests {
         gen_graph::models::insert_change(&conn, &change).unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &block_group_id, false)
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &block_group_id, false)
                 .unwrap(),
             HashSet::from_iter([
                 "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
@@ -287,7 +287,7 @@ mod tests {
         gen_graph::models::insert_change(&conn, &change).unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(&conn, &block_group_id, false)
+            gen_graph::models::get_all_sequences_with_pruning(&conn, &block_group_id, false)
                 .unwrap(),
             HashSet::from_iter([
                 "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
@@ -353,7 +353,7 @@ mod tests {
         );
         let block_group = get_sample_bg(conn, &collection, "derived");
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(conn, &block_group.id, false)
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_group.id, false)
                 .unwrap(),
             HashSet::from_iter([
                 "ATCGATCGATCGATCGATCGGGAACACACAGAGA".to_string(),
@@ -414,12 +414,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
         assert_eq!(
@@ -532,12 +528,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -595,12 +587,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -664,12 +652,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -727,12 +711,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -790,12 +770,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
     }
@@ -842,12 +818,8 @@ mod tests {
         );
         assert_eq!(block_groups.len(), 1);
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
-                conn,
-                &block_groups[0].id,
-                false
-            )
-            .unwrap(),
+            gen_graph::models::get_all_sequences_with_pruning(conn, &block_groups[0].id, false)
+                .unwrap(),
             HashSet::from_iter(expected_sequences),
         );
