+++ src/updates/vcf.rs
@@ -780,7 +780,7 @@ mod tests {
             false,
         )?;
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, Sample::DEFAULT_NAME).id,
                 false,
@@ -790,7 +790,7 @@ mod tests {
         );
         // `G1` genotype has no changes
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, "G1").id,
                 false
@@ -800,7 +800,7 @@ mod tests {
         );
         // `foo` is homozygous for the first variant and does not contain the second
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, "foo").id,
                 false
@@ -840,7 +840,7 @@ mod tests {
         )
         .unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, Sample::DEFAULT_NAME).id,
                 false,
@@ -850,7 +850,7 @@ mod tests {
         );
         // `bar` sample has the refrence + a deletion of the C
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, "bar").id,
                 false
@@ -863,7 +863,7 @@ mod tests {
         );
         // `baz` sample has a deletion of CG and an insertion of A
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, "baz").id,
                 false
@@ -905,7 +905,7 @@ mod tests {
         )
         .unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, Sample::DEFAULT_NAME).id,
                 false,
@@ -914,7 +914,7 @@ mod tests {
             HashSet::from_iter(vec!["ATCGATCGATCGATCGATCGGGAACACACAGAGA".to_string()])
         );
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, "sample 1").id,
                 false
@@ -963,7 +963,7 @@ mod tests {
         .unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, "sample 1").id,
                 false
@@ -1038,7 +1038,7 @@ mod tests {
         .unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, "unknown").id,
                 false
@@ -1081,7 +1081,7 @@ mod tests {
         )
         .unwrap();
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, "foo").id,
                 false
@@ -1125,7 +1125,7 @@ mod tests {
         .unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, "foo").id,
                 true
@@ -1451,7 +1451,7 @@ mod tests {
         .unwrap();
 
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, Sample::DEFAULT_NAME).id,
                 true,
@@ -1460,7 +1460,7 @@ mod tests {
             HashSet::from_iter(vec!["ATCGATCGATCGATCGATCGGGAACACACAGAGA".to_string()])
         );
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, "f1").id,
                 true
@@ -1469,7 +1469,7 @@ mod tests {
             HashSet::from_iter(vec!["ATCTCGATCGATCGCGGGAACACACAGAGA".to_string()])
         );
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, "f2").id,
                 true
@@ -1478,7 +1478,7 @@ mod tests {
             HashSet::from_iter(vec!["ATCTGGATCGATCGCGGAATCAGAACACACAGGA".to_string()])
         );
         assert_eq!(
-            gen_models_graph_tests::get_all_sequences_with_pruning(
+            gen_graph::models::get_all_sequences_with_pruning(
                 conn,
                 &get_sample_bg(conn, &collection, "f3").id,
                 true
@@ -1527,7 +1527,7 @@ mod tests {
         )
         .unwrap();
 
-        let child_sequences = gen_models_graph_tests::get_all_sequences_with_pruning(
+        let child_sequences = gen_graph::models::get_all_sequences_with_pruning(
             conn,
             &get_sample_bg(conn, &collection, "child").id,
             true,
@@ -1583,7 +1583,7 @@ mod tests {
         )
         .unwrap();
 
-        let child_sequences = gen_models_graph_tests::get_all_sequences_with_pruning(
+        let child_sequences = gen_graph::models::get_all_sequences_with_pruning(
             conn,
             &get_sample_bg(conn, &collection, "child").id,
             true,
